--- kmp/arch/metro-retain/src/commonMain/kotlin/io/github/reactivecircus/kstreamlined/kmp/arch/metro/retain/Retainable.kt
@@ -1,10 +1,0 @@
-package io.github.reactivecircus.kstreamlined.kmp.arch.metro.retain
-
-/**
- * Marker interface for types that can be retained via [metroRetain].
- *
- * Used as the bound type for [RetainedKey] map contributions.
- * Implement this on any class that should be injectable and retainable
- * through the Metro + retain integration.
- */
-public interface Retainable
